(#808) Group Swagger endpoints into sections
